rename provisioner sample to deployer and factor out code that will be common to a firefox version
