+++ src/cloud/deployer/freedom-module.json
@@ -1,0 +1,23 @@
+{
+  "name": "deployer",
+  "description": "Deploy a uProxy cloud instance.",
+  "app": {
+    "script": [
+      "freedom-module.static.js"
+    ]
+  },
+  "dependencies": {
+    "loggingcontroller": {
+      "url": "../../loggingprovider/freedom-module.json",
+      "api": "loggingcontroller"
+    },
+    "logginglistener": {
+      "url": "../../loggingprovider/freedom-module.json",
+      "api": "logginglistener"
+    },
+    "digitalocean": {
+      "url": "../digitalocean/freedom-module.json",
+      "api": "digitalocean"
+    }
+  }
+}
