--- src/samples/provision-chromeapp/index.html
@@ -1,8 +1,0 @@
-<html>
-  <head>
-    <title>provision</title>
-  </head>
-  <body>
-    <h1>provision</h1>
-  </body>
-</html>
